change: relocate scaffold
